Refactor CompiledQuerySourceBuilder.walkReferencedAssemblies into utility class
